backend/requirements.txt
fastapi
uvicorn
duckdb
pandas
python-dotenv
google-generativeai
plotly
